Exercise #5, flex modal
Used flex to recreate a modal window. Learned more about how containers
work. The modal was the container for all the content. Then we had two
flex items, the icon and the container of the header, text, and buttons.
The header and close button were moved together and flexed to achieve
the space between them on the window. The rest of the items in the
container were not flexed, but remained in block format.